% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{ggp2bpGetBoxPlotAtmm}
\alias{ggp2bpGetBoxPlotAtmm}
\title{Boxplot of TMM data overlayed with points of control milk yields}
\usage{
ggp2bpGetBoxPlotAtmm(ptdfResult, psMonthNr, ptdfControlMilkYield)
}
\arguments{
\item{ptdfResult}{tbl_df of transferred TMM data}

\item{psMonthNr}{Month formatted as '%m'}

\item{ptdfControlMilkYield}{tbl_df of control milk yields}
}
\value{
ggp2bpResult ggplot object containing overlayed plots
}
\description{
Given a tbl_df with transferred TMM data and a tbl_df of control milk yields,
a boxplot is generated using the TMM data and the control milk yields are
overlayed as red-colored points.
}
