% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{tdfGetControlMilkYield}
\alias{tdfGetControlMilkYield}
\title{Convert a dataframe of control milk yield to tbl_df}
\usage{
tdfGetControlMilkYield(pdfControlMilkYield, pdfResult)
}
\arguments{
\item{pdfControlMilkYield}{original dataframe of control milk yield}

\item{pdfResult}{original dataframe of TMM records}
}
\value{
tdfControlMilkYield extended tbl_df of control milk yield
}
\description{
The dataframe of control milk yields is converted to a tbl_df.
A mapping to cow indices of the TMM records given in pdfResult
is generated and added to the records of control milk yield
}
