% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{tdfGetTmmPerMonth}
\alias{tdfGetTmmPerMonth}
\title{Extend transferred daily milk with cow index}
\usage{
tdfGetTmmPerMonth(pdfResult, psMonth)
}
\arguments{
\item{pdfResult}{original data.frame containing all data transferred from database}

\item{psMonth}{Month (formatted as %m) for which results should be returned}
}
\value{
tdfTmmPerMonth tbl_df containing resulting milk records from month psMonth
}
\description{
Daily milk yield is transferred from the database as
a dataframe. The data frame has basically daily milk
yields (TMM) for each cows over a period of months.
The result of the function \code{tdfGetTmmPerMonth}
is a tbl_df{dplyr} of all milk records for the
given month psMonth for all cows.
}
\details{
The resulting tbl_df has an additional column with
the name ATMM_KUH_IDX. This column has indices that
represent the identities of the cow. This index is
only used for plotting to have more readable axis
labels.
}
