% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{vecCleanUpChar}
\alias{vecCleanUpChar}
\title{Remove fixed set of characters from given vector of strings}
\usage{
vecCleanUpChar(pvecRemoveChar, pvecSrc)
}
\arguments{
\item{pvecRemoveChar}{vector of characters to be removed}

\item{pvecSrc}{vector of strings from which characters should be removed}
}
\value{
vecCleanResult vector of strings w/out removed characters
}
\description{
Given a vector of characters to be removed from a vector of strings,
this is done with a loop over the characters to be removed using
fixed version of gsub. This could be done easier with regular
expressions
}
