% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{vecGetTmmLifeNr}
\alias{vecGetTmmLifeNr}
\title{Extract unique cow lifenumbers from complete TMM dataset}
\usage{
vecGetTmmLifeNr(pdfResult)
}
\arguments{
\item{pdfResult}{original data frame of TMM records}
}
\value{
vecTmmLifeNr vector of unique cow life numbers found in pdfResult
}
\description{
Given a complete set of transferred milk yield records (TMM), all
unique cow lifenumbers are extracted and returned as a vector
}
